split out plotting fans
